+++ src/macros/clonable_fixed_alias.rs
@@ -100,12 +100,12 @@ macro_rules! cloneable_fixed_alias {
             }
 
             #[cfg(feature = "encoding-bech32")]
-            fn try_to_bech32(&self, hrp: &str) -> Result<alloc::string::String, $crate::Bech32EncodingError> {
+            fn try_to_bech32(&self, hrp: &str) -> Result<alloc::string::String, $crate::Bech32Error> {
                 self.expose_secret().try_to_bech32(hrp)
             }
 
             #[cfg(feature = "encoding-bech32")]
-            fn try_to_bech32m(&self, hrp: &str) -> Result<alloc::string::String, $crate::Bech32EncodingError> {
+            fn try_to_bech32m(&self, hrp: &str) -> Result<alloc::string::String, $crate::Bech32Error> {
                 self.expose_secret().try_to_bech32m(hrp)
             }
         }