+++ src/macros/serializable_dynamic_alias.rs
@@ -79,12 +79,12 @@ macro_rules! serializable_dynamic_alias {
             }
 
             #[cfg(feature = "encoding-bech32")]
-            fn try_to_bech32(&self, hrp: &str) -> Result<alloc::string::String, $crate::Bech32EncodingError> {
+            fn try_to_bech32(&self, hrp: &str) -> Result<alloc::string::String, $crate::Bech32Error> {
                 $crate::ExposeSecret::expose_secret(self).try_to_bech32(hrp)
             }
 
             #[cfg(feature = "encoding-bech32")]
-            fn try_to_bech32m(&self, hrp: &str) -> Result<alloc::string::String, $crate::Bech32EncodingError> {
+            fn try_to_bech32m(&self, hrp: &str) -> Result<alloc::string::String, $crate::Bech32Error> {
                 $crate::ExposeSecret::expose_secret(self).try_to_bech32m(hrp)
             }
         }